Complete refactoring
